Add files via upload
Worked on another way to do the movement (I dit every direction as a separate State, this was bad). I also added a start for the falling stone. It moves on contact, collides with player etc. , but doesn't kill the player yet. And I added the pump you can throw, it can not interact with enemies as there are no enemies yet.